Split out override checker into its own file.
